dashboard functionality complete
