+++ lib/initializerData.ts
@@ -1,0 +1,14 @@
+export const initialIntolerances = [
+  { name: 'dairy', avoid: false },
+  { name: 'egg', avoid: false },
+  { name: 'gluten', avoid: false },
+  { name: 'grain', avoid: false },
+  { name: 'peanut', avoid: false },
+  { name: 'seafood', avoid: false },
+  { name: 'sesame', avoid: false },
+  { name: 'shellfish', avoid: false },
+  { name: 'soy', avoid: false },
+  { name: 'sulfite', avoid: false },
+  { name: 'tree_nut', avoid: false },
+  { name: 'wheat', avoid: false },
+];
