+++ pages/index.js
@@ -1,7 +1,7 @@
 import { useUserContext } from '@/contexts/UserContext';
 
 export default function Home() {
-  const { user } = useUserContext();
+  const { user, username } = useUserContext();
 
   return (
     <div className='p-4 shadow rounded bg-white'>
