+++ tailwind.config.js
@@ -12,6 +12,13 @@ module.exports = {
       '3/4': '75%',
       full: '100%',
     },
+    maxWidth: {
+      0: '0',
+      '1/4': '25%',
+      '1/2': '50%',
+      '3/4': '75%',
+      full: '100%',
+    },
     extend: {
       outline: {
         green: '3px solid rgb(16, 185, 129)',
